+++ HISTORY.md
@@ -13,6 +13,10 @@
 - Option 'oph_fastdelete'
 - Options to refer to substrings of the value of workflow variables
 
+### Changed:
+
+- Pid of output cube of the tasks skipped because of errors is set to pid of input cube
+
 ## v1.9.1 - 2025-01-04
 
 ### Fixed: